+++ .gitignore
@@ -28,3 +28,5 @@ yarn-error.log*
 api/venv
 api/__pycache__/
 venv
+
+.gradle
