--- src/TechLabs.js
@@ -1,40 +1,0 @@
-
-
-
-import React,
-    class TechLabs extends
-        constructor() {
-            super()
-                this.state = {
-                     recipes: [],}}
-                         componentDidMount() {
-                            fetch("
-                            .then(response =
-                        .then(recipes => t
-                    render() {
-                 return (
-                <div>
-             <h1>Recipes</h1>{
-           <div>
-        this.state.recipes.
-       <p>{entry["weekday"]}:
-    ))</div>}
-</div>);}
-};
-                                export
-                                default
-                                TechLabs;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
