+++ src/index.js
@@ -4,12 +4,12 @@ import reportWebVitals from './reportWebVitals';
 import 'semantic-ui-css/semantic.min.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App.js';
-import './testApp.css';
-import './recipeForm.css';
-import './App.css';
-import './login.css';
-import './index.css';
-import './startpage.css';
+import '../css/testApp.css';
+import '../css/recipeForm.css';
+import '../css/App.css';
+import '../css/login.css';
+import '../css/index.css';
+import '../css/startpage.css';
 import './showIngredients';
 
 ReactDOM.render(
@@ -17,8 +17,4 @@ ReactDOM.render(
     document.getElementById('root')
 );
 
-
-///If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
